fix some gcc warnings
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@2590 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/DiscIO/Src/Blob.cpp
@@ -71,7 +71,7 @@ bool SectorReader::Read(u64 offset, u64 size, u8* out_ptr)
 	while (remain > 0)
 	{
 		// Check if we are ready to do a large block read. > instead of >= so we don't bother if remain is only one block.
-		if (positionInBlock == 0 && remain > m_blocksize)
+		if (positionInBlock == 0 && remain > (u64)m_blocksize)
 		{
 			u64 num_blocks = remain / m_blocksize;
 			ReadMultipleAlignedBlocks(block, num_blocks, out_ptr);
@@ -110,7 +110,7 @@ bool SectorReader::Read(u64 offset, u64 size, u8* out_ptr)
 
 bool SectorReader::ReadMultipleAlignedBlocks(u64 block_num, u64 num_blocks, u8 *out_ptr)
 {
-	for (int i = 0; i < num_blocks; i++)
+	for (u64 i = 0; i < num_blocks; i++)
 	{
 		const u8 *data = GetBlockData(block_num + i);
 		if (!data)
